requirements.txt

Django
Pillow
reportlab
django-crispy-forms
python-dotenv
six>=1.16.0


asgiref
certifi
charset-normalizer

google-api-core
google-auth
google-cloud-core
google-cloud-vision
google.generativeai
googleapis-common-protos
idna
protobuf
pyasn1
pyasn1-modules
python-dateutil

requests
rsa
six
sqlparse
typing_extensions
uritemplate
urllib3
gunicorn
django-jazzmin
matplotlib
seaborn
whitenoise
